requirements.txt
pygame-ce==2.5.2
pytest==7.4.2